considering microdata
Although in this case, I don't believe the microdata would be strictly
necessary because this kind of information would generally only be
accessible behind some sort of internally authenticated system.

As such, the specific itemScope or itemType attributes might not be as
helpful as something that is more domain specific, or something that
reflects the nuances of how an organization might structure their
information.